delete unit test that doesn't really make sense anymore
